+++ src/packageA/function/MathUtil.java
@@ -2,7 +2,7 @@ package packageA.function;
 
 public class MathUtil {
 	
-	public static Object solve(int left, char op, int right) {
+	public static Object solve(float left, char op, float right) {
 		switch (op) {
 	        case '*': return left * right;
 	        case '/': return left / right;
